Break out intro content into Welcome component, add tests'
