+++ src/components/MainNav/__tests__/MainNav.test.js
@@ -1,0 +1,8 @@
+import { render, screen } from "@testing-library/react";
+import MainNav from "../MainNav";
+
+test('renders MainNav element', () => {
+  render(<MainNav />);
+  const mainNavElement = screen.getByRole('banner');
+  expect(mainNavElement).toBeInTheDocument();
+});
