+++ src/components/Welcome/WelcomeStyles.js
@@ -10,11 +10,7 @@ const useStyles = makeStyles((theme) => ({
   },
   buttons: {
     marginTop: "40px",
-  },
-  footer: {
-    backgroundColor: theme.palette.background.paper,
-    padding: "50px 0",
-  },
+  }
 }));
 
 export default useStyles;
